Implemented answer page

+++ PoligraphApp/Review+CoreDataClass.swift
@@ -17,7 +17,7 @@ public class Review: NSManagedObject {
             Includes all info to fully define review, no other attributes need to be added.
  
     */
-    class func addNewReview(feedbackText: String, recommendation: Int, biasRating: Double, question: Question, reviewer: User, inManagedObjectContext context: NSManagedObjectContext) -> Review? {
+    class func addNewReview(feedbackText: String, recommendation: Int, biasRating: Float, question: Question, reviewer: User, inManagedObjectContext context: NSManagedObjectContext) -> Review? {
         
         if let review = NSEntityDescription.insertNewObject(forEntityName: "Review", into: context) as? Review {
             review.feedbackText = feedbackText
